let mktime figure out if DST is in effect
Since the tm structure is always zeroed out before calling strptime(),
the tm_isdst field would always be 0. The value 0 has semantic meaning
for mktime(), i.e. DST is not in effect. A negative value on the other
hand instruct mktime() to figure out if DST is in effect. Discovered
while runnning the test suite from Sweden on March 31 where DST is in
full swing.