add support for using any character as the pattern delimiter
By default, slash (`/') is the expected pattern delimiter. Instead
support any character as the pattern delimiter, similar to how sed(1)
behaves. This is especially useful when one wants to match a literal
slash inside a pattern; removing the need to escape it:

  $ cat mdsort.conf
  maildir "~/Maildir/INBOX" {
    match header "Subject" |/sys/| label "sys"
  }

Implemented using a lexer backdoor. A concept described in the original
"Yacc: Yet Another Compiler Compiler" paper.