remove duplicate label logic
The fact labels where protected against adding duplicates is implicit
and not documented anywhere. Therefore remove the functionality
completely instead, leading to fewer surprises.

This also exposed a problem with the label tests. A message would get
moved from the new to the src directory and then moved again. In order
to prevent this, a global blacklist functionality must be implemented in
order to only operate on a message once. Don't know if this is worth
doing right now.