blacklist messages moved to the current maildir
If a message ends up being moved to the maildir that's currently being
traversed, blacklist the same message in order to prevent operating on
it more than once. This does not affect moving messages to another
maildir.

The inode of the message is used to determine if the message has already
been seen; that's hopefully stable enough.

The blacklist functionality is written in a way to avoid overhead for
the common operation of moving messages to another maildir.