reject config consisting of only a pass action
Consider the following configuration:

  maildir "~/Maildir/INBOX" {
    match header "From" /example/ pass
  }

... performing a dry-run would cause an assertion to trigger in
matches_inspect() since no non-NULL path is found. This is caused by the
fact that pass matches are removed in expr_eval_block() in order to
continue evaluation. This in turn leaves a single match header left.

The "easiest" solution is just to reject such configuration since it's a
NOP anyway.