correct dangling pass logic
The attempt to fix this problem is commit ef41dcc ("reject config
consisting of only a pass action") was not sufficient.

A pass action could potentially not be followed by an effective action.
Determining this while parsing the configuration is not possible as it
depends on the actual message and if it ends up being matched by a
following expression. Instead detect this at runtime and treat it as a
no match. This ensures matches_inspect() isn't called without a
destination path.