add exec action
This action can be used to execute an arbitrary command for each matched
message. Multiple arguments can be expressed using a string list, which
is an existing grammar construct.

The command to execute is consciously passed to execvp(2) and not
`sh -c' as the command may be interpolated with user supplied data.
This is of course no guarantee from allowing potentially dangerous
configurations, see the tests in tests/action-exec.sh.