skip mbox separator line
While reading a message from stdin, a mbox separator line could be
present before the actual message. I was recently hit by this while
running a fairly old version of OpenSMTPD 6.0.3 on Linux while
delivering local using mdsort as a MDA.

The concept of a mbox separator line is defined in RFC 4155 under the
"About the mbox Database " section.