stop requiring explicit new lines in mdsort.conf
The purpose of requiring explicit new lines after some grammar
constructs was to simplify error handling as it allowed the current line
to be discarded. This is a best effort approach allowing the parser to
hopefully continue, in the hopes of reporting more errors instead of
halting.

Instead, stop requiring explicit new lines and skip the current line
while encountering an error. This makes line continuations (i.e.
backslash) redundant and is therefore a breaking change.

Existing configurations must be adjusted as follows:

	$ cat mdsort.conf
	maildir "~/Maildir/INBOX" {
		match date > 2 weeks \
			move "~/Maildir/Archive"
	}
	$ sed -i -e 's/[[:space:]]*\\//' mdsort.conf
	$ cat mdsort.conf
	maildir "~/Maildir/INBOX" {
		match date > 2 weeks
			move "~/Maildir/Archive"
	}