make label changes visible to exec
Adding a label to a message followed by an exec action did not make the
label changes visible to the exec command. The label action would write
out a new message with the label applied but the message still kept the
old file descriptor open, referring to the outdated message at this
point.

Instead, update the message file descriptor in maildir_write(). Some
fiddling is needed as we want a readable file descriptor.